requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

# API Clients
openai==1.3.7
deepgram-sdk==3.0.0
httpx==0.25.2

# WebSocket Support
websockets==12.0
python-socketio==5.10.0

# Data Storage
redis==5.0.1

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
